refactor(dev): manifest utils for compiler
